requirements.txt
# Agentic Insight Analyst - Python Dependencies

# Core Dependencies
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Text Processing & NLP
scikit-learn>=1.3.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
rich>=13.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
